Refactor authentication related code and introduce new models
This commit includes refactoring of authentication related code, along
with the introduction of new models for user authentication and
authorization. The authentication process and user roles have been
restructured, and new models such as `UserModel`, `Role`, and `Token`
have been introduced to handle user authentication and authorization.
Additionally, a new base model for location-based features has been
added. The changes aim to improve the maintainability and extensibility
of the authentication module. No issues are associated with this commit.

Affected files:
- Deleted: lib/application/auth/auth_bloc.dart
- Deleted: lib/application/auth/auth_bloc.freezed.dart
- Deleted: lib/application/auth/auth_event.dart
- Deleted: lib/application/auth/auth_state.dart
- Added: lib/domain/common/enums/role.dart
- Added: lib/domain/repositories/auth_service_repo.dart
- Added: lib/infrastructure/implementations/auth/auth_service.dart
- Added: lib/infrastructure/models/auth/token.dart (deleted)
- Added: lib/infrastructure/models/auth/token.freezed.dart (deleted)
- Added: lib/infrastructure/models/auth/token.g.dart (deleted)
- Added: lib/infrastructure/models/auth/user_model.dart
- Added: lib/infrastructure/models/home/core/base/base_model.dart

This commit does not fix any specific issues but represents a
significant refactoring and expansion of the authentication module.